% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkR-package.R
\docType{package}
\name{BulkR-package}
\alias{BulkR}
\alias{BulkR-package}
\title{BulkR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
