% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{combine_results}
\alias{combine_results}
\title{Title}
\usage{
combine_results(
  ...,
  filter = TRUE,
  fdr = NULL,
  l2fc = NULL,
  type = "list",
  significant_genes = "all"
)
}
\arguments{
\item{significant_genes}{}
}
\description{
Title
}
