% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_counts.R
\name{expected_counts}
\alias{expected_counts}
\title{Title}
\usage{
expected_counts(counts, factor = 1e+06, log = TRUE, pseudocount = 1)
}
\arguments{
\item{counts}{A counts matrix}

\item{factor}{Integer, a scaling factor}

\item{log}{Logical, log counts?}
}
\value{
An integer of expected counts
}
\description{
Title
}
\examples{
expected_counts(Seurat@assays$RNA@counts)
expected_counts(Seurat@assays$RNA@counts, factor=1e3, log=TRUE, pseudocount=0.5)
}
