% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_L2FC_scatter.R
\name{plot_L2FC_scatter}
\alias{plot_L2FC_scatter}
\title{Title}
\usage{
plot_L2FC_scatter(
  DE_Results_comb,
  x = NULL,
  y = NULL,
  genes = "All",
  flavor = "Min",
  fdr = 0.05,
  l2fc = 0,
  symmetric = TRUE,
  guides = TRUE,
  pt.size = 1,
  annotate_genes = NULL
)
}
\arguments{
\item{DE_results_comb}{}
}
\description{
Title
}
