% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_comb_DE_results.R
\name{select_comb_DE_results}
\alias{select_comb_DE_results}
\title{Title}
\usage{
select_comb_DE_results(DE_Results_comb, X = NULL, Y = NULL, type = "L2FC")
}
\arguments{
\item{DE_results_comb}{}
}
\description{
Title
}
