% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial_sva_DESeq2.R
\name{serial_sva}
\alias{serial_sva}
\title{Serial SVA DESeq2 analysis}
\usage{
serial_sva(DDS, n.sv = NULL, parallel = TRUE)
}
\arguments{
\item{parallel}{}
}
\description{
Serial SVA DESeq2 analysis
}
