% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout_fraq.R
\name{dropout_fraq}
\alias{dropout_fraq}
\title{Title}
\usage{
dropout_fraq(counts, expected_counts)
}
\arguments{
\item{counts}{Numeric vector, counts of a single cell/sample}

\item{expected_counts}{Numeric vector, expected counts per gene/feature in the same order as counts}
}
\value{
dropout fraction per expected_counts bin
}
\description{
Title
}
\examples{
dropout_fraq(Seurat@assays$RNA@counts[1,], expected_counts(Seurat@assays$RNA@counts))
}
