% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_counts.R
\name{expected_counts}
\alias{expected_counts}
\title{Title}
\usage{
expected_counts(counts, factor = 1e+06)
}
\arguments{
\item{factor}{}
}
\value{
An integer of expected counts
}
\description{
Title
}
\examples{
expected_counts(Seurat@assays$RNA@counts)
}
